Make the index names configurable using env vars
