+++ api/catalog/api/constants/media_types.py
@@ -5,4 +5,6 @@ Also see `ingestion_server/constants/media_types.py`
 AUDIO_TYPE = "audio"
 IMAGE_TYPE = "image"
 
-MEDIA_TYPES = [(AUDIO_TYPE, "Audio"), (IMAGE_TYPE, "Image")]
+MEDIA_TYPES = [AUDIO_TYPE, IMAGE_TYPE]
+
+MEDIA_TYPE_CHOICES = [(AUDIO_TYPE, "Audio"), (IMAGE_TYPE, "Image")]
