+++ api/catalog/configuration/aws.py
@@ -1,0 +1,5 @@
+from decouple import config
+
+
+AWS_ACCESS_KEY_ID = config("AWS_ACCESS_KEY_ID", default="")
+AWS_SECRET_ACCESS_KEY = config("AWS_SECRET_ACCESS_KEY", default="")
