+++ api/env.template
@@ -41,6 +41,9 @@ SEMANTIC_VERSION="1.0.0"
 #ELASTICSEARCH_PORT="9200"
 #ELASTICSEARCH_AWS_REGION="us-east-1"
 
+#IMAGE_INDEX_NAME="image"
+#AUDIO_INDEX_NAME="audio"
+
 #AWS_ACCESS_KEY_ID=""
 #AWS_SECRET_ACCESS_KEY=""
